% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equaldis.TStest.HD.r
\name{Equaldis.TStest.HD}
\alias{Equaldis.TStest.HD}
\title{A two-sample test for the equality of distributions for high-dimensional data}
\usage{
Equaldis.TStest.HD(X, Y, method = c("spect", "spect_ind", "boot", "us",
  "us_ind", "perm"))
}
\arguments{
\item{X}{A matrix where each row is one of the p-samples in the first group.}

\item{Y}{A matrix where each row is one of the p-samples in the second group.}

\item{method}{the two-sample test. By default the “us” method is computed. See details.}
}
\value{
A list containing the following components:
\item{standarized statistic: }{the value of the standarized statistic.}
\item{p.value: }{the p-value for the test.}
\item{statistic: }{the value of the statistic.}
\item{variance: }{the value of the variance estimator.}
\item{p: }{number of samples or populations.}
\item{n: }{sample size in the first group.}
\item{m: }{sample size in the second group.}
\item{method: }{a character string indicating which two sample test is performed.}
\item{I.statistics: }{the p individual statistics.}
\item{I.permutation.p.values: }{the p individual permutation p-values.}
\item{data.name: }{a character string giving the name of the data.}
}
\description{
A two-sample test for the equality of distributions for high-dimensional data
}
\details{
The function implements the two-sample tests proposed by Cousido-Rocha, et al. (2018). The methods “spect”,“boot” and “us” are based on a global statistic which is the average of individual statistics
corresponding to each of the p variables. Each of these individual statistics measure the difference between
the empirical characteristic functions computed from the two samples. An alternative expression of them
show that it can be interpreted as a difference between the variability in each of the two samples and the
variability in the combined sample. The global statistic (average) is standarized using different variance
estimators given place to the three different methods. The method “spect” uses a variance estimator
based on the spectral analysis theory, the method “boot” implements the block bootstrap to estimate the
variance and the method “us” employs a variance estimators derived from U-statistic theory
(more details in Cousido-Rocha et al., 2018). The methods “spect” and “boot” are suitable under some theoretical
assumptions which include that the sequence of individual statistics that defined the global statistic is
strictly stationary whereas the method “us” avoids such assumption. However the methods “spect” and
“boot” have been checked in simulations and they perform well even when such assumption is violated.
The asymptotic normality (when p tends to infinity) of the standardized version of the statistic is used
to compute the corresponding p-value. On the other hand, Cousido-Rocha et al. (2018) also proposed
the method “perm” whose global statistic is the avergae of the permutation p-values corresponding to
the individual statistics mentioned above. This method assumes that the sequence of p-values is strictly
stationary, however in simulations it seems that it performs where this assumption does not hold.
Furthermore than defining a new global test these p-values can be also used when the global null hypothesis.
}
\references{
Cousido-Rocha, M., de Uña-Álvarez J., and Hart, J. (2018). A two-sample test for the equality of distributions for high-dimensional data. Preprint.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{José Carlos Soage González}
\item{Jacobo de Uña-Álvarez}
\item{D. Hart, Jeffrey.}
}
}
