% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.r
\docType{data}
\name{p}
\alias{p}
\title{Equaldis.TStest.HD}
\format{An object of class \code{numeric} of length 1.}
\usage{
p
}
\description{
A two-sample test for the equality of distributions for high-dimensional data
}
\details{
The function implements the two-sample tests proposed by Cousido-Rocha, et al. (2018). The methods “spect”,“boot” and “us” are based on a global statistic which is the average of individual statistics
corresponding to each of the p variables. Each of these individual statistics measure the difference between
the empirical characteristic functions computed from the two samples. An alternative expression of them
show that it can be interpreted as a difference between the variability in each of the two samples and the
variability in the combined sample. The global statistic (average) is standarized using different variance
estimators given place to the three different methods. The method “spect” uses a variance estimator
based on the spectral analysis theory, the method “boot” implements the block bootstrap to estimate the
variance and the method “us” employs a variance estimators derived from U-statistic theory
(more details in Cousido-Rocha et al., 2018). The methods “spect” and “boot” are suitable under some theoretical
assumptions which include that the sequence of individual statistics that defined the global statistic is
strictly stationary whereas the method “us” avoids such assumption. However the methods “spect” and
“boot” have been checked in simulations and they perform well even when such assumption is violated.
The asymptotic normality (when p tends to infinity) of the standardized version of the statistic is used
to compute the corresponding p-value. On the other hand, Cousido-Rocha et al. (2018) also proposed
the method “perm” whose global statistic is the avergae of the permutation p-values corresponding to
the individual statistics mentioned above. This method assumes that the sequence of p-values is strictly
stationary, however in simulations it seems that it performs where this assumption does not hold.
Furthermore than defining a new global test these p-values can be also used when the global null hypothesis.
}
\keyword{datasets}
